Refactoring LOss struct to losses module
