+++ delta/src/neuralnet/models/sequential.rs
@@ -34,9 +34,9 @@ use std::path::Path;
 use std::time::Instant;
 
 use crate::common::layer::Layer;
-use crate::common::loss::Loss;
 use crate::common::{ModelError, Tensor};
 use crate::dataset::{Dataset, ImageDatasetOps};
+use crate::losses::Loss;
 use crate::optimizers::Optimizer;
 
 /// A sequential model that contains a list of layers, an optimizer, and a loss function.
